andromeda: timer: Rework the irq system for timer

+++ include/types.h
@@ -111,7 +111,7 @@ typedef signed char int8_t;
 // typedef struct __TYPE_REGISTERS isrVal_t;
 typedef unsigned long size_t;
 typedef unsigned long long idx_t;
-typedef long time_t; // Time variables are specified to be signed
+typedef signed long long time_t; // Time variables are specified to be signed
 
 typedef unsigned long addr_t;
 typedef long reg;
